third_party/WebKit/LayoutTests/plugins/type-case.html

<!DOCYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<embed id="p1" src="resources/test.blinktestplugin" type="application/x-blin&#x212A;-test-plugin">
<object id="p2" src="resources/test.blinktestplugin" type="application/x-blin&#x212A;-test-plugin">
<script>
test(() => {
  assert_equals(document.querySelector('#p1').postMessage, undefined);
  assert_equals(document.querySelector('#p2').postMessage, undefined);
}, 'MIME type matching should be ASCII case-insensitive.');
</script>
